requirements.txt
langchain
langchain_community
openai
pandas
flask
tabulate
keras
langchain_openai
python-dotenv
openpyxl
streamlit
pandasai
pypdf
tensorflow
tiktoken
sentence-transformers
faiss-cpu